+++ src/main/java/ru/vova777/coordinatePlane/SecondCoordinateQuarter.java
@@ -19,7 +19,7 @@ public class SecondCoordinateQuarter extends CoordinateQuarter {
 
     @Override
     public double getAzimuth(double delta_x, double lengthTrack) {
-        double angle = Math.toDegrees(Math.acos((double) delta_x / lengthTrack));
+        double angle = Math.toDegrees(Math.acos(delta_x / lengthTrack));
         return (180 - angle);
     }
